requirement_core.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.40.26
botocore==1.40.26
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
databricks-sdk==0.65.0
debugpy==1.8.16
decorator==5.2.1
docker==7.1.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
nest-asyncio==1.6.0
nlpaug==1.1.11
nltk==3.9.1
numpy==2.2.6
nvidia-nccl-cu12==2.28.3
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.9.1
requests==2.32.5
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.3
wordcloud==1.9.4
xgboost==3.0.5
zipp==3.23.0
